Spelling & Formatting
Most of the changes here are correcting the spelling of Zernike, as well as fixing several small spelling errors in examples so that they will run as designed. One of the changes includes fixing Zernike spelling in a filename, so it looks as if the file's been deleted and readded under another name.